+++ src/file_io.c
@@ -8,7 +8,7 @@
 const char *FILE_EXTENSIONS[] = {".csv", ".bin"};
 #define NUM_EXTENSIONS (sizeof(FILE_EXTENSIONS) / sizeof(FILE_EXTENSIONS[0]))
 
-float **import_file(const char *file_name, size_t *num_rows, size_t *num_cols)
+double **import_file(const char *file_name, size_t *num_rows, size_t *num_cols)
 {
 
 	char file_path[200];
@@ -24,7 +24,7 @@ float **import_file(const char *file_name, size_t *num_rows, size_t *num_cols)
 	}
 
 	size_t capacity = INITIAL_CAPACITY;
-	float **data = (float **)malloc(capacity * sizeof(float *));
+	double **data = (double **)malloc(capacity * sizeof(double *));
 	if (!data)
 	{
 		printf("Error: Memory allocation failed.\n");
@@ -48,7 +48,7 @@ float **import_file(const char *file_name, size_t *num_rows, size_t *num_cols)
 		if (row >= capacity)
 		{
 			capacity *= 2;
-			float **temp = (float **)realloc(data, capacity * sizeof(float *));
+			double **temp = (double **)realloc(data, capacity * sizeof(double *));
 			if (!temp)
 			{
 				printf("Error: Memory reallocation failed.\n");
@@ -62,7 +62,7 @@ float **import_file(const char *file_name, size_t *num_rows, size_t *num_cols)
 		char *token = strtok(line, ","); // Split by comma
 
 		// Allocate memory for columns (start with 100, expand dynamically)
-		data[row] = (float *)malloc(100 * sizeof(float));
+		data[row] = (double *)malloc(100 * sizeof(double));
 		size_t col_capacity = 100;
 
 		while (token != NULL)
@@ -70,7 +70,7 @@ float **import_file(const char *file_name, size_t *num_rows, size_t *num_cols)
 			if (col >= col_capacity)
 			{
 				col_capacity *= 2;
-				float *temp_row = (float *)realloc(data[row], col_capacity * sizeof(float));
+				double *temp_row = (double *)realloc(data[row], col_capacity * sizeof(double));
 				if (!temp_row)
 				{
 					printf("Error: Column reallocation failed.\n");
@@ -102,7 +102,7 @@ float **import_file(const char *file_name, size_t *num_rows, size_t *num_cols)
 }
 
 // Function to print the full dataset in a table format
-void print_data(float **data, size_t num_rows, size_t num_cols)
+void print_data(double **data, size_t num_rows, size_t num_cols)
 {
 	if (!data)
 	{
@@ -135,11 +135,11 @@ void print_data(float **data, size_t num_rows, size_t num_cols)
 	printf("--------------------------------------------------------------\n");
 }
 
-int verify_data(float **data, size_t data_num_rows, size_t data_num_cols, const char *file)
+int verify_data(double **data, size_t data_num_rows, size_t data_num_cols, const char *file)
 {
 	size_t verify_num_rows = 0; // Variable to store the number of rows read
 	size_t verify_num_cols = 0; // Variable to store the number of columns read
-	float **verify_data = import_file(file, &verify_num_rows, &verify_num_cols);
+	double **verify_data = import_file(file, &verify_num_rows, &verify_num_cols);
 
 	if (!verify_data)
 	{
